Crate corelib_rust

Re-exports

pub use retrieve::ky;

pub use retrieve::proxied;

pub use retrieve::unlimited;

pub use markets::nasdaq::api_nasdaq_quotes::nasdaq_quote;

pub use markets::nasdaq::api_nasdaq_quotes::nasdaq_quotes;

pub use markets::nasdaq::api_nasdaq_quotes::AssetClass;

pub use markets::nasdaq::api_nasdaq_unlimited::get_nasdaq_headers;

pub use markets::nasdaq::api_nasdaq_unlimited::nasdaq_end_point;

pub use markets::nasdaq::api_nasdaq_unlimited::nasdaq_end_points;

pub use markets::nasdaq::datafeeds::streaming::yahoo::EventRecord;

pub use markets::nasdaq::datafeeds::streaming::yahoo::LogRecord;

pub use markets::nasdaq::datafeeds::streaming::yahoo::RustCallbacks;

pub use markets::nasdaq::datafeeds::streaming::yahoo::YahooConfig;

pub use markets::nasdaq::datafeeds::streaming::yahoo::YahooStreaming;

pub use markets::nasdaq::datafeeds::streaming::yahoo::YahooStreamingCore;

Modules

markets

Public markets module containing Nasdaq and Yahoo data feed logic.

retrieve

Public retrieve module (contains the ky HTTP client).

utils

Public utils module (mirrors TS utils section). Contains internal helpers for future FFI functions (cron, etc.).

Functions

get_version

Returns the Cargo package version as a string. Used by the TS FFI bridge (getVersion).

log_and_double

Logs a message in Rust and returns the input value doubled. Used by the TS FFI bridge (logAndDouble).